+++ src/interpreter/mod.rs
@@ -1,5 +1,6 @@
 pub mod env;
 pub mod expr;
+pub mod types;
 use env::{Env, FuncKind, VarKind};
 use expr::Expression;
 pub mod load_std;